git commit -m "Add dynamic project name to Flask route
This commit includes changes to the Flask route to fetch the project name from the database and pass it to the Jinja2 template. This allows the project name to be displayed dynamically on the webpage.

TODO: Fix rendering issues in 'base.html' and 'all_projects.html, at the moment there is no list of all projects rendering once a project is selected'."